refactor: extract auth crate from codex-core
