+++ codex-rs/auth/src/storage_tests.rs
