+++ codex-rs/core/src/lib.rs
@@ -104,6 +104,8 @@ pub type CodexConversation = CodexThread;
 pub use analytics_client::AnalyticsEventsClient;
 pub use auth::AuthManager;
 pub use auth::CodexAuth;
+pub use auth::from_auth_storage_with_core_client;
+pub use auth::load_auth_with_core_client;
 pub mod default_client;
 pub mod project_doc;
 mod rollout;
