+++ codex-rs/core/src/util.rs
@@ -4,7 +4,6 @@ use std::time::Duration;
 
 use codex_protocol::ThreadId;
 use rand::Rng;
-use tracing::debug;
 use tracing::error;
 
 use crate::parse_command::shlex_join;
@@ -50,21 +49,6 @@ pub(crate) fn error_or_panic(message: impl std::string::ToString) {
     } else {
         error!("{}", message.to_string());
     }
-}
-
-pub(crate) fn try_parse_error_message(text: &str) -> String {
-    debug!("Parsing server error response: {}", text);
-    let json = serde_json::from_str::<serde_json::Value>(text).unwrap_or_default();
-    if let Some(error) = json.get("error")
-        && let Some(message) = error.get("message")
-        && let Some(message_str) = message.as_str()
-    {
-        return message_str.to_string();
-    }
-    if text.is_empty() {
-        return "Unknown error".to_string();
-    }
-    text.to_string()
 }
 
 pub fn resolve_path(base: &Path, path: &PathBuf) -> PathBuf {
