+++ codex-rs/core/src/util_tests.rs
@@ -1,30 +1,6 @@
 use super::*;
 
 #[test]
-fn test_try_parse_error_message() {
-    let text = r#"{
-  "error": {
-    "message": "Your refresh token has already been used to generate a new access token. Please try signing in again.",
-    "type": "invalid_request_error",
-    "param": null,
-    "code": "refresh_token_reused"
-  }
-}"#;
-    let message = try_parse_error_message(text);
-    assert_eq!(
-        message,
-        "Your refresh token has already been used to generate a new access token. Please try signing in again."
-    );
-}
-
-#[test]
-fn test_try_parse_error_message_no_error() {
-    let text = r#"{"message": "test"}"#;
-    let message = try_parse_error_message(text);
-    assert_eq!(message, r#"{"message": "test"}"#);
-}
-
-#[test]
 fn feedback_tags_macro_compiles() {
     #[derive(Debug)]
     struct OnlyDebug;
